tests: Add separate WSI test file
